+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [9.25.0] - 2024-09-29
 ### Added
 - **feat**: `null` as a rule to validate that a value is null, take note this can be used in tandem with conditional groups, for example:
 ```typescript